requirements.txt
flask==3.0.2
pdfminer.six==20231228
markdown==3.6
langchain==0.1.14
langchain-openai==0.1.1
unstructured==0.11.8
python-dotenv==1.0.1
openai==1.16.2
chromadb==0.4.24
tiktoken==0.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
boto3==1.34.79
qdrant-client
qdrant-client[fastembed]
fastembed